backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
openai==1.3.0
google-generativeai==0.3.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
werkzeug==2.3.7
